Changement du prix des unit
int en double

+++ include/Unite.h
@@ -15,7 +15,7 @@ class Unite
     protected:
 
     private:
-        unsigned int prix;
+        unsigned double prix;
         unsigned int pv;
         unsigned int pa;
 };
